rename unions files to groups
