+++ src/components/screens/Groups/Groups.stories.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { storiesOf } from "@storybook/react";
+
+import MyComponent from "../Groups";
+
+storiesOf("Screens", module).add("Groups", () => <MyComponent />);
