+++ src/components/screens/Groups/Groups.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import MyComponent from "../Unions";
+import MyComponent from "../Groups";
 
 it("renders without crashing", () => {
   shallow(<MyComponent />);
