--- src/components/screens/Unions/Unions.stories.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-import { storiesOf } from "@storybook/react";
-
-import MyComponent from "../Unions";
-
-storiesOf("Screens", module).add("Unions", () => <MyComponent />);
